front-chat/requirements.txt
streamlit>=1.15.0
langchain
langchain-openai
python-dotenv
yfinance>=0.2.0
pandas>=1.3.0
plotly>=5.3.0
numpy>=1.21.0
ta>=0.10.0
langgraph>=0.0.17
pydantic>=1.8.0
fpdf2>=2.7.0
secure-smtplib>=0.1.0
email-validator>=1.1.0
kaleido>=0.2.0
fastapi>=0.68.0
uvicorn>=0.15.0
twilio>=7.0.0
python-multipart>=0.0.5
requests>=2.26.0
pydub>=0.25.0
SpeechRecognition>=3.8.0
streamlit-webrtc>=0.45.0
av>=9.2.0
authlib>=1.0.0
PyJWT>=2.0.0
extra-streamlit-components>=0.1.0
google-auth-oauthlib>=0.4.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0